+++ client/client.py
@@ -1,5 +1,3 @@
-import click
-from socket import *
 import time
 import json
 
@@ -52,22 +50,3 @@ class Client:
     def action_handler(self, action, **kwargs):
         if action == 'probe':
             return self.presence()
-
-
-@click.command()
-@click.argument('address', default="localhost")
-@click.argument('port', default=7777)
-def start(address, port):
-    user = Client('ivanov', '123', 'online')
-    with socket(AF_INET, SOCK_STREAM) as s:  # Создает сокет TCP
-        s.connect((address, port))  # Присваивает адрес и порт
-        s.send(user.authenticate())
-        tm = s.recv(MAX_MSG_SIZE)
-        print(json.loads(tm.decode(ENCODING)))
-        s.send(user.disconnect())
-        # tm = s.recv(MAX_MSG_SIZE)
-        # print(tm.decode(ENCODING))
-
-
-if __name__ == '__main__':
-    start()
